Refactored application to organized folders and files.
