+++ client/components/Application/Sections/Languages.js
@@ -1,0 +1,17 @@
+import React from 'react';
+import MultipleComplex from '../Inputs/MultipleComplexInput';
+
+function Languages() {
+  return (
+    <MultipleComplex mainName={'Languages'} subInputs={[
+      {
+        name: 'languages_language', schemaName: 'language', displayName: 'Language:', placeholder: '', parent: 'languages',
+      },
+      {
+        name: 'languages_fluency', schemaName: 'fluency', displayName: 'Fluency:', placeholder: '', parent: 'languages',
+      },
+    ]} />
+  );
+}
+
+export default Languages;
