+++ client/components/Application/Sections/index.js
@@ -1,0 +1,27 @@
+import Awards from './Awards';
+import Basics from './Basics';
+import Education from './Education';
+import Interests from './Interests';
+import Keywords from './Keywords';
+import Languages from './Languages';
+import PublishedWork from './PublishedWork';
+import References from './References';
+import Skills from './Skills';
+import Volunteer from './Volunteer';
+import Work from './Work';
+
+const index = {
+  Awards,
+  Basics,
+  Education,
+  Interests,
+  Keywords,
+  Languages,
+  PublishedWork,
+  References,
+  Skills,
+  Volunteer,
+  Work,
+};
+
+export default index;
